+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - (Feature) Init Operator Auth Secret
 - (Feature) (Platform) Share gateway resources
 - (Bugfix) (Platform) Escape URL in Meta Integration
+- (Feature) Unify Webhooks
 
 ## [1.3.4](https://github.com/arangodb/kube-arangodb/tree/1.3.4) (2025-12-22)
 - (Feature) Add scrape annotations for ArangoD pods